+++ src/main/java/com/lambdaschool/crudyorders/OrdersappApplication.java
@@ -1,0 +1,13 @@
+package com.lambdaschool.crudyorders;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrdersappApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OrdersappApplication.class, args);
+	}
+
+}
