+++ src/main/java/com/lambdaschool/crudyorders/models/Agent.java
@@ -1,0 +1,104 @@
+package com.lambdaschool.crudyorders.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "agents")
+public class Agent
+{
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(nullable = false)
+    private long agentcode;
+
+    private String agentname;
+    private String workingarea;
+    private double commission;
+    private String phone;
+    private String country;
+
+    // relationship to Customer
+    @OneToMany(mappedBy = "agentcode", cascade = CascadeType.ALL, orphanRemoval = true)
+    @JsonIgnoreProperties(value = "agentcode")
+    private List<Customer> customers = new ArrayList<>();
+
+    // default constructor
+    public Agent()
+    {
+    }
+
+    //constructor
+    // (agentcode, agentname, workingarea, commission, phone, country)
+    public Agent(String agentname, String workingarea, double commission, String phone, String country)
+    {
+        this.agentcode = agentcode;
+        this.agentname = agentname;
+        this.workingarea = workingarea;
+        this.commission = commission;
+        this.phone = phone;
+        this.country = country;
+    }
+
+    // getters and setters
+    public long getAgentcode() {
+        return agentcode;
+    }
+
+    public void setAgentcode(long agentcode) {
+        this.agentcode = agentcode;
+    }
+
+    public String getAgentname() {
+        return agentname;
+    }
+
+    public void setAgentname(String agentname) {
+        this.agentname = agentname;
+    }
+
+    public String getWorkingarea() {
+        return workingarea;
+    }
+
+    public void setWorkingarea(String workingarea) {
+        this.workingarea = workingarea;
+    }
+
+    public double getCommission() {
+        return commission;
+    }
+
+    public void setCommission(double commission) {
+        this.commission = commission;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+    public String getCountry() {
+        return country;
+    }
+
+    public void setCountry(String country) {
+        this.country = country;
+    }
+
+    public List<Customer> getCustomers() {
+        return customers;
+    }
+
+    public void setCustomers(List<Customer> customers) {
+        this.customers = customers;
+    }
+
+
+}
