+++ src/main/java/com/lambdaschool/crudyorders/models/Customer.java
@@ -1,0 +1,187 @@
+package com.lambdaschool.crudyorders.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "customers")
+@JsonIgnoreProperties(value = "hasOpeningAmount") // internal use only must be ignored
+public class Customer
+{
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(nullable = false)
+    private long custcode;
+
+    @Column(nullable = false)
+    private String custname;
+    private String custcity;
+    private String workingarea;
+    private String custcountry;
+    private String grade;
+
+    @Transient
+    public boolean hasOpeningAmount = false; // declaring variable hasOpeningAmount
+    private double openingamt;
+
+    @Transient
+    public boolean hasReceiveAmount = false; // declaring variable hasReceiveAmount
+    private double receiveamt;
+
+
+    @Transient
+    public boolean hasPaymentAmount = false; // declaring variable hasPaymentAmount
+    private double paymentamt;
+
+    @Transient
+    public boolean hasOutstandingAmount = false; // declaring variable hasOutstandingAmount
+    private double outstandingamt;
+
+    private String phone;
+
+    // foreign key relationship to Agent
+    @ManyToOne
+    @JoinColumn(name = "agentcode", nullable = false)
+    @JsonIgnoreProperties(value = "customers")
+    private Agent agentcode;
+
+    //relationship to Order
+    @OneToMany(mappedBy = "custcode", cascade = CascadeType.ALL, orphanRemoval = true)
+    @JsonIgnoreProperties(value = "custcode")
+    private List<Order> orders = new ArrayList<>();
+
+    //default constructor
+    public Customer() {
+    }
+
+    //constructor
+    // (custcode, custname, custcity, workingarea, custcountry, grade, openingamt, receiveamt, paymentamt, outstandingamt, phone, agentcode)
+    public Customer(String custname, String custcity, String workingarea, String custcountry, String grade, double openingamt, double receiveamt, double paymentamt, double outstandingamt, String phone, Agent agentcode) {
+        this.custcode = custcode;
+        this.custname = custname;
+        this.custcity = custcity;
+        this.workingarea = workingarea;
+        this.custcountry = custcountry;
+        this.grade = grade;
+        this.openingamt = openingamt;
+        this.receiveamt = receiveamt;
+        this.paymentamt = paymentamt;
+        this.outstandingamt = outstandingamt;
+        this.phone = phone;
+        this.agentcode = agentcode;
+    }
+
+    // getters and setters
+    public long getCustcode() {
+        return custcode;
+    }
+
+    public void setCustcode(long custcode) {
+        this.custcode = custcode;
+    }
+
+    public String getCustname() {
+        return custname;
+    }
+
+    public void setCustname(String custname) {
+        this.custname = custname;
+    }
+
+    public String getCustcity() {
+        return custcity;
+    }
+
+    public void setCustcity(String custcity) {
+        this.custcity = custcity;
+    }
+
+    public String getWorkingarea() {
+        return workingarea;
+    }
+
+    public void setWorkingarea(String workingarea) {
+        this.workingarea = workingarea;
+    }
+
+    public String getCustcountry() {
+        return custcountry;
+    }
+
+    public void setCustcountry(String custcountry) {
+        this.custcountry = custcountry;
+    }
+
+    public String getGrade() {
+        return grade;
+    }
+
+    public void setGrade(String grade) {
+        this.grade = grade;
+    }
+
+    public double getOpeningamt() {
+        return openingamt;
+    }
+
+    public void setOpeningamt(double openingamt)
+    {
+        hasOpeningAmount = true; // checker
+        this.openingamt = openingamt;
+    }
+
+    public double getReceiveamt() { return receiveamt; }
+
+    public void setReceiveamt(double receiveamt)
+    {
+        hasReceiveAmount = true; // checker
+        this.receiveamt = receiveamt;
+    }
+
+    public double getPaymentamt() {
+        return paymentamt;
+    }
+
+    public void setPaymentamt(double paymentamt)
+    {
+        hasPaymentAmount = true; // checker
+        this.paymentamt = paymentamt;
+    }
+
+    public double getOutstandingamt() {
+        return outstandingamt;
+    }
+
+    public void setOutstandingamt(double outstandingamt)
+    {
+        hasOutstandingAmount = true; // checker
+        this.outstandingamt = outstandingamt;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+    public Agent getAgentcode() {
+        return agentcode;
+    }
+
+    public void setAgentcode(Agent agentcode) {
+        this.agentcode = agentcode;
+    }
+
+    public List<Order> getOrders() {
+        return orders;
+    }
+
+    public void setOrders(List<Order> orders) {
+        this.orders = orders;
+    }
+}
