+++ src/main/java/com/lambdaschool/crudyorders/repositories/AgentsRepository.java
@@ -1,0 +1,10 @@
+package com.lambdaschool.crudyorders.repositories;
+
+import com.lambdaschool.crudyorders.models.Agent;
+import org.springframework.data.repository.CrudRepository;
+
+public interface AgentsRepository extends CrudRepository<Agent, Long>
+{
+    Agent findByAgentcode(long agentcode);
+
+}
