+++ src/main/java/com/lambdaschool/crudyorders/repositories/OrdersRepository.java
@@ -1,0 +1,11 @@
+package com.lambdaschool.crudyorders.repositories;
+
+import com.lambdaschool.crudyorders.models.Order;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
+public interface OrdersRepository extends CrudRepository<Order, Long>
+{
+    List<Order> findByAdvanceamountGreaterThan(double amount);
+}
