+++ src/main/java/com/lambdaschool/crudyorders/services/AgentServices.java
@@ -1,0 +1,11 @@
+package com.lambdaschool.crudyorders.services;
+
+import com.lambdaschool.crudyorders.models.Agent;
+
+public interface AgentServices
+{
+    Agent save(Agent agent);
+
+    Agent findByAgentcode(long agentcode);
+
+}
