+++ src/main/java/com/lambdaschool/crudyorders/services/PaymentServices.java
@@ -1,0 +1,8 @@
+package com.lambdaschool.crudyorders.services;
+
+import com.lambdaschool.crudyorders.models.Payment;
+
+public interface PaymentServices
+{
+    Payment save(Payment payment);
+}
