+++ src/main/java/com/lambdaschool/crudyorders/view/OrderCount.java
@@ -1,0 +1,7 @@
+package com.lambdaschool.crudyorders.view;
+
+public interface OrderCount
+{
+    String getCustname();
+    int getOrdercount();
+}
