--- src/main/java/com/lambdaschool/ordersapp/OrdersappApplication.java
@@ -1,13 +1,0 @@
-package com.lambdaschool.ordersapp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrdersappApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrdersappApplication.class, args);
-	}
-
-}
