--- src/main/java/com/lambdaschool/ordersapp/services/AgentServices.java
@@ -1,11 +1,0 @@
-package com.lambdaschool.ordersapp.services;
-
-import com.lambdaschool.ordersapp.models.Agent;
-
-public interface AgentServices
-{
-    Agent save(Agent agent);
-
-    Agent findByAgentcode(long agentcode);
-
-}
