--- src/main/java/com/lambdaschool/ordersapp/services/CustomerServices.java
@@ -1,24 +1,0 @@
-package com.lambdaschool.ordersapp.services;
-
-import com.lambdaschool.ordersapp.models.Customer;
-import com.lambdaschool.ordersapp.view.OrderCount;
-
-import java.util.List;
-
-public interface CustomerServices
-{
-    Customer save(Customer customer);
-
-    List<Customer> findCustomerOrders();
-
-    Customer findByCustCode(long custcode);
-
-    List<Customer> findByCustomerName(String subname);
-
-    List<OrderCount> countOrderByCustomer();
-
-    Customer update(Customer customer, long custcode);
-
-    void delete(long custcode);
-
-}
