--- src/main/java/com/lambdaschool/ordersapp/services/OrderServices.java
@@ -1,19 +1,0 @@
-package com.lambdaschool.ordersapp.services;
-
-import com.lambdaschool.ordersapp.models.Order;
-
-import java.util.List;
-
-public interface OrderServices
-{
-    Order save(Order order);
-
-    Order findOrderByOrdnum(long ordnum);
-
-    List<Order> getByAdvanceamount(double amount);
-
-    List<Order> findAllOrders();
-
-    void delete(long ordernum);
-
-}
