--- src/main/java/com/lambdaschool/ordersapp/services/PaymentServices.java
@@ -1,8 +1,0 @@
-package com.lambdaschool.ordersapp.services;
-
-import com.lambdaschool.ordersapp.models.Payment;
-
-public interface PaymentServices
-{
-    Payment save(Payment payment);
-}
