--- src/main/java/com/lambdaschool/ordersapp/view/OrderCount.java
@@ -1,7 +1,0 @@
-package com.lambdaschool.ordersapp.view;
-
-public interface OrderCount
-{
-    String getCustname();
-    int getOrdercount();
-}
