+++ src/test/java/com/lambdaschool/crudyorders/OrdersappApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lambdaschool.crudyorders;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrdersappApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
