--- src/test/java/com/lambdaschool/ordersapp/OrdersappApplicationTests.java
@@ -1,13 +1,0 @@
-package com.lambdaschool.ordersapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrdersappApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
